requirements.txt
python-docx
Pillow
opencv-python
numpy
PyQt5
customtkinter
